fix(vm): http_server do_call no-cache sentinel u16::MAX -> u32::MAX
Straggler from the inline-cache cid widening (1756746f, u16 -> u32).
After widening, do_call's cache_id is u32 and 65535 (u16::MAX) is a
REAL cache slot, no longer the "no cache" sentinel — so this call site
in install_rack_input (the rack-input StringIO.new) was caching into
slot 65535 instead of bypassing the cache. Use u32::MAX, the actual
no-cache sentinel.

+++ crates/rubyrs/src/http_server.rs
@@ -546,7 +546,7 @@ fn install_rack_input(
     let pre_frames = pg.vm.frames.len();
     pg.vm.stack.push(sio_class);
     pg.vm.stack.push(body_val);
-    pg.vm.do_call(new_sym, 1, /* no_recv = */ false, /* cache_id = */ u16::MAX)?;
+    pg.vm.do_call(new_sym, 1, /* no_recv = */ false, /* cache_id = */ u32::MAX)?;
     // Drive `initialize` to completion. `new` uses the frame's
     // `swap_return` so the value left on the stack is the INSTANCE, not
     // initialize's return value.
